requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
python-dotenv==1.0.1
mem0ai>=1.0.0
qdrant-client>=1.9.1
openai>=1.0.0
